refactor(backend): split god-files into cohesive siblings [Phase 6]
Pure, behavior-preserving file splits within the same packages (Go doesn't care which
file a method lives in) to improve navigability.

- game.go (2351 -> 1912): extract the VP recalculation context + VP methods into
  game_vp.go, and the event-subscription setup (subscribeToVPEvents / generational /
  ocean / global-parameter) into game_subscriptions.go.
- state_calculator.go (2126 -> 1706): extract milestone/award state into
  state_calculator_achievements.go, choice validation into state_calculator_choices.go,
  and behavior/colony value computation into state_calculator_behavior_values.go.

No symbols renamed, no API changes. Verified: go build ./... exit 0; golangci-lint 0
issues; gofmt clean; go test ./test/... 26 packages, 0 failures.
