+++ backend/internal/action/state_calculator.go
@@ -6,10 +6,8 @@ import (
 	"time"
 
 	"terraforming-mars-backend/internal/game"
-	"terraforming-mars-backend/internal/game/award"
 	gamecards "terraforming-mars-backend/internal/game/cards"
 	"terraforming-mars-backend/internal/game/global_parameters"
-	"terraforming-mars-backend/internal/game/milestone"
 	"terraforming-mars-backend/internal/game/player"
 	"terraforming-mars-backend/internal/game/shared"
 )
@@ -1705,422 +1703,4 @@ func formatInsufficientTagsMessage(tag string) string {
 
 func formatTooManyTagsMessage(tag string) string {
 	return fmt.Sprintf("Too many %s tags", formatTagDisplayName(tag))
-}
-
-// CalculateMilestoneState computes eligibility state for claiming a milestone.
-// Returns EntityState with errors indicating why the milestone cannot be claimed.
-func CalculateMilestoneState(
-	milestoneType shared.MilestoneType,
-	p *player.Player,
-	g *game.Game,
-	cardRegistry gamecards.CardRegistry,
-	milestoneRegistry milestone.MilestoneRegistry,
-) player.EntityState {
-	var errors []player.StateError
-	metadata := make(map[string]interface{})
-
-	def, err := milestoneRegistry.GetByID(string(milestoneType))
-	if err != nil {
-		errors = append(errors, player.StateError{
-			Code:     player.ErrorCodeInvalidRequirement,
-			Category: player.ErrorCategoryConfiguration,
-			Message:  fmt.Sprintf("Unknown milestone type: %s", milestoneType),
-		})
-		return player.EntityState{
-			Errors:         errors,
-			Cost:           make(map[string]int),
-			Metadata:       metadata,
-			LastCalculated: time.Now(),
-		}
-	}
-
-	ms := g.Milestones()
-
-	errors = append(errors, validateActionsRemaining(p, g)...)
-	errors = append(errors, validateNoPendingSelection(p, g)...)
-
-	if ms.IsClaimed(milestoneType) {
-		errors = append(errors, player.StateError{
-			Code:     player.ErrorCodeMilestoneAlreadyClaimed,
-			Category: player.ErrorCategoryAchievement,
-			Message:  "Already claimed",
-		})
-	}
-
-	if ms.ClaimedCount() >= game.MaxClaimedMilestones {
-		errors = append(errors, player.StateError{
-			Code:     player.ErrorCodeMaxMilestonesClaimed,
-			Category: player.ErrorCategoryAchievement,
-			Message:  "Maximum milestones claimed",
-		})
-	}
-
-	progress := gamecards.CalculateMilestoneProgress(def, p, g.Board(), cardRegistry)
-	required := def.GetRequired()
-	metadata["progress"] = progress
-	metadata["required"] = required
-
-	if progress < required {
-		errors = append(errors, player.StateError{
-			Code:     player.ErrorCodeMilestoneRequirementNotMet,
-			Category: player.ErrorCategoryRequirement,
-			Message:  "Requirement not met",
-		})
-	}
-
-	cost := def.ClaimCost
-	if p.Resources().Get().Credits < cost {
-		errors = append(errors, player.StateError{
-			Code:     player.ErrorCodeInsufficientCredits,
-			Category: player.ErrorCategoryCost,
-			Message:  "Cannot afford",
-		})
-	}
-
-	costMap := map[string]int{string(shared.ResourceCredit): cost}
-
-	return player.EntityState{
-		Errors:         errors,
-		Cost:           costMap,
-		Metadata:       metadata,
-		LastCalculated: time.Now(),
-	}
-}
-
-// CalculateAwardState computes eligibility state for funding an award.
-// Returns EntityState with errors indicating why the award cannot be funded.
-func CalculateAwardState(
-	awardType shared.AwardType,
-	p *player.Player,
-	g *game.Game,
-	awardRegistry award.AwardRegistry,
-) player.EntityState {
-	var errors []player.StateError
-	metadata := make(map[string]interface{})
-
-	def, err := awardRegistry.GetByID(string(awardType))
-	if err != nil {
-		errors = append(errors, player.StateError{
-			Code:     player.ErrorCodeInvalidRequirement,
-			Category: player.ErrorCategoryConfiguration,
-			Message:  fmt.Sprintf("Unknown award type: %s", awardType),
-		})
-		return player.EntityState{
-			Errors:         errors,
-			Cost:           make(map[string]int),
-			Metadata:       metadata,
-			LastCalculated: time.Now(),
-		}
-	}
-
-	gameAwards := g.Awards()
-
-	errors = append(errors, validateActionsRemaining(p, g)...)
-	errors = append(errors, validateNoPendingSelection(p, g)...)
-
-	if gameAwards.IsFunded(awardType) {
-		errors = append(errors, player.StateError{
-			Code:     player.ErrorCodeAwardAlreadyFunded,
-			Category: player.ErrorCategoryAchievement,
-			Message:  "Already funded",
-		})
-	}
-
-	if gameAwards.FundedCount() >= game.MaxFundedAwards {
-		errors = append(errors, player.StateError{
-			Code:     player.ErrorCodeMaxAwardsFunded,
-			Category: player.ErrorCategoryAchievement,
-			Message:  "Maximum awards funded",
-		})
-	}
-
-	cost := def.GetCostForFundedCount(gameAwards.FundedCount())
-	metadata["fundingCost"] = cost
-
-	if p.Resources().Get().Credits < cost {
-		errors = append(errors, player.StateError{
-			Code:     player.ErrorCodeInsufficientCredits,
-			Category: player.ErrorCategoryCost,
-			Message:  "Cannot afford",
-		})
-	}
-
-	costMap := map[string]int{string(shared.ResourceCredit): cost}
-
-	return player.EntityState{
-		Errors:         errors,
-		Cost:           costMap,
-		Metadata:       metadata,
-		LastCalculated: time.Now(),
-	}
-}
-
-// CalculateChoiceErrors validates a single choice's requirements against the player/game state.
-// Returns a list of errors explaining why the choice is unavailable (empty if available).
-func CalculateChoiceErrors(choice shared.Choice, p *player.Player, g *game.Game, cardRegistry gamecards.CardRegistry) []player.StateError {
-	var errors []player.StateError
-
-	if choice.Requirements != nil && len(choice.Requirements.Items) > 0 {
-		for _, req := range choice.Requirements.Items {
-			if err := checkChoiceRequirement(req, p, g, cardRegistry); err != nil {
-				errors = append(errors, *err)
-			}
-		}
-	}
-
-	resources := p.Resources().Get()
-	for _, outputBC := range choice.Outputs {
-		if shared.IsVariableAmount(outputBC) || outputBC.GetAmount() >= 0 {
-			continue
-		}
-		if !isBasicPlayerResource(outputBC.GetResourceType()) {
-			continue
-		}
-		available := resources.GetAmount(outputBC.GetResourceType())
-		if available < -outputBC.GetAmount() {
-			errors = append(errors, player.StateError{
-				Code:     player.ErrorCodeInsufficientResources,
-				Category: player.ErrorCategoryInput,
-				Message:  fmt.Sprintf("Not enough %s", outputBC.GetResourceType()),
-			})
-		}
-	}
-
-	for _, inputBC := range choice.Inputs {
-		if shared.IsVariableAmount(inputBC) || inputBC.GetAmount() <= 0 {
-			continue
-		}
-		if !isBasicPlayerResource(inputBC.GetResourceType()) {
-			continue
-		}
-		available := resources.GetAmount(inputBC.GetResourceType())
-		if available < inputBC.GetAmount() {
-			errors = append(errors, player.StateError{
-				Code:     player.ErrorCodeInsufficientResources,
-				Category: player.ErrorCategoryInput,
-				Message:  fmt.Sprintf("Not enough %s", inputBC.GetResourceType()),
-			})
-		}
-	}
-
-	return errors
-}
-
-// checkChoiceRequirement validates a single choice requirement and returns a StateError if not met.
-func checkChoiceRequirement(req shared.ChoiceRequirement, p *player.Player, g *game.Game, cardRegistry gamecards.CardRegistry) *player.StateError {
-	switch req.Type {
-	case "tags":
-		if req.Tag == nil {
-			return &player.StateError{
-				Code:     player.ErrorCodeInvalidRequirement,
-				Category: player.ErrorCategoryRequirement,
-				Message:  "Invalid tag requirement",
-			}
-		}
-		tagCount := gamecards.CountPlayerTagsByType(p, cardRegistry, *req.Tag)
-		if req.Min != nil && tagCount < *req.Min {
-			return &player.StateError{
-				Code:     player.ErrorCodeInsufficientTags,
-				Category: player.ErrorCategoryRequirement,
-				Message:  formatInsufficientTagsMessage(string(*req.Tag)),
-			}
-		}
-		if req.Max != nil && tagCount > *req.Max {
-			return &player.StateError{
-				Code:     player.ErrorCodeTooManyTags,
-				Category: player.ErrorCategoryRequirement,
-				Message:  formatTooManyTagsMessage(string(*req.Tag)),
-			}
-		}
-
-	case "temperature":
-		temp := g.GlobalParameters().Temperature()
-		if req.Min != nil && temp < *req.Min {
-			return &player.StateError{
-				Code:     player.ErrorCodeTemperatureTooLow,
-				Category: player.ErrorCategoryRequirement,
-				Message:  "Temperature too low",
-			}
-		}
-		if req.Max != nil && temp > *req.Max {
-			return &player.StateError{
-				Code:     player.ErrorCodeTemperatureTooHigh,
-				Category: player.ErrorCategoryRequirement,
-				Message:  "Temperature too high",
-			}
-		}
-
-	case "oxygen":
-		oxygen := g.GlobalParameters().Oxygen()
-		if req.Min != nil && oxygen < *req.Min {
-			return &player.StateError{
-				Code:     player.ErrorCodeOxygenTooLow,
-				Category: player.ErrorCategoryRequirement,
-				Message:  "Oxygen too low",
-			}
-		}
-		if req.Max != nil && oxygen > *req.Max {
-			return &player.StateError{
-				Code:     player.ErrorCodeOxygenTooHigh,
-				Category: player.ErrorCategoryRequirement,
-				Message:  "Oxygen too high",
-			}
-		}
-
-	case "ocean":
-		oceans := g.GlobalParameters().Oceans()
-		if req.Min != nil && oceans < *req.Min {
-			return &player.StateError{
-				Code:     player.ErrorCodeOceansTooLow,
-				Category: player.ErrorCategoryRequirement,
-				Message:  "Too few oceans",
-			}
-		}
-		if req.Max != nil && oceans > *req.Max {
-			return &player.StateError{
-				Code:     player.ErrorCodeOceansTooHigh,
-				Category: player.ErrorCategoryRequirement,
-				Message:  "Too many oceans",
-			}
-		}
-
-	case "tr":
-		tr := p.Resources().TerraformRating()
-		if req.Min != nil && tr < *req.Min {
-			return &player.StateError{
-				Code:     player.ErrorCodeTRTooLow,
-				Category: player.ErrorCategoryRequirement,
-				Message:  "TR too low",
-			}
-		}
-		if req.Max != nil && tr > *req.Max {
-			return &player.StateError{
-				Code:     player.ErrorCodeTRTooHigh,
-				Category: player.ErrorCategoryRequirement,
-				Message:  "TR too high",
-			}
-		}
-
-	case "production":
-		if req.Resource == nil {
-			return &player.StateError{
-				Code:     player.ErrorCodeInvalidRequirement,
-				Category: player.ErrorCategoryRequirement,
-				Message:  "Invalid production requirement",
-			}
-		}
-		production := p.Resources().Production()
-		amount := production.GetAmount(*req.Resource)
-		if req.Min != nil && amount < *req.Min {
-			return &player.StateError{
-				Code:     player.ErrorCodeInsufficientProduction,
-				Category: player.ErrorCategoryRequirement,
-				Message:  formatInsufficientProductionMessage(*req.Resource),
-			}
-		}
-
-	case "resource":
-		if req.Resource == nil {
-			return &player.StateError{
-				Code:     player.ErrorCodeInvalidRequirement,
-				Category: player.ErrorCategoryRequirement,
-				Message:  "Invalid resource requirement",
-			}
-		}
-		resources := p.Resources().Get()
-		amount := resources.GetAmount(*req.Resource)
-		if req.Min != nil && amount < *req.Min {
-			return &player.StateError{
-				Code:     player.ErrorCodeInsufficientResources,
-				Category: player.ErrorCategoryRequirement,
-				Message:  formatInsufficientResourceMessage(*req.Resource),
-			}
-		}
-	}
-
-	return nil
-}
-
-// computeBehaviorValues computes per-condition output values for card behaviors.
-// Returns a slice of ComputedBehaviorValue with target format "behaviors::N".
-func computeBehaviorValues(
-	behaviors []shared.CardBehavior,
-	sourceCardID string,
-	p *player.Player,
-	g *game.Game,
-	cardRegistry gamecards.CardRegistry,
-	colonyBonusLookup gamecards.ColonyBonusLookup,
-) []player.ComputedBehaviorValue {
-	var result []player.ComputedBehaviorValue
-
-	board := g.Board()
-	allPlayers := g.GetAllPlayers()
-
-	for i, behavior := range behaviors {
-		var outputs []shared.CalculatedOutput
-		for _, outputBC := range behavior.Outputs {
-			if outputBC.GetResourceType() == shared.ResourceColonyBonus {
-				bonusOutputs := computeColonyBonusValues(p, g, colonyBonusLookup)
-				if outputs == nil {
-					outputs = bonusOutputs
-				} else {
-					outputs = append(outputs, bonusOutputs...)
-				}
-				continue
-			}
-			per := shared.GetPerCondition(outputBC)
-			if per == nil {
-				continue
-			}
-			count := gamecards.CountPerCondition(per, sourceCardID, p, board, cardRegistry, allPlayers)
-			if per.Amount > 0 {
-				multiplier := count / per.Amount
-				actualAmount := outputBC.GetAmount() * multiplier
-				outputs = append(outputs, shared.CalculatedOutput{
-					ResourceType: string(outputBC.GetResourceType()),
-					Amount:       actualAmount,
-					IsScaled:     true,
-				})
-			}
-		}
-		for _, choice := range behavior.Choices {
-			for _, outputBC := range choice.Outputs {
-				per := shared.GetPerCondition(outputBC)
-				if per == nil {
-					continue
-				}
-				count := gamecards.CountPerCondition(per, sourceCardID, p, board, cardRegistry, allPlayers)
-				if per.Amount > 0 {
-					multiplier := count / per.Amount
-					actualAmount := outputBC.GetAmount() * multiplier
-					outputs = append(outputs, shared.CalculatedOutput{
-						ResourceType: string(outputBC.GetResourceType()),
-						Amount:       actualAmount,
-						IsScaled:     true,
-					})
-				}
-			}
-		}
-		if outputs != nil {
-			result = append(result, player.ComputedBehaviorValue{
-				Target:  fmt.Sprintf("behaviors::%d", i),
-				Outputs: outputs,
-			})
-		}
-	}
-
-	return result
-}
-
-func computeColonyBonusValues(
-	p *player.Player,
-	g *game.Game,
-	colonyBonusLookup gamecards.ColonyBonusLookup,
-) []shared.CalculatedOutput {
-	if p == nil || g == nil || colonyBonusLookup == nil || !g.HasColonies() {
-		return []shared.CalculatedOutput{}
-	}
-	return gamecards.ColonyBonusesToCalculatedOutputs(
-		gamecards.CollectColonyBonuses(p.ID(), g.Colonies().States(), colonyBonusLookup),
-	)
 }
