+++ backend/internal/game/game.go
@@ -21,156 +21,6 @@ import (
 	"terraforming-mars-backend/internal/game/shared"
 )
 
-type VPCardInfo struct {
-	CardID       string
-	CardName     string
-	CardType     string
-	Description  string
-	VPConditions []shared.VPCondition
-	Tags         []shared.CardTag
-}
-
-type VPCardLookup interface {
-	LookupVPCard(cardID string) (*VPCardInfo, error)
-}
-
-type gameVPRecalculationContext struct {
-	game *Game
-}
-
-func (ctx *gameVPRecalculationContext) GetCardStorage(playerID string, cardID string) int {
-	p, err := ctx.game.GetPlayer(playerID)
-	if err != nil {
-		return 0
-	}
-	return p.Resources().GetCardStorage(cardID)
-}
-
-func (ctx *gameVPRecalculationContext) CountPlayerTagsByType(playerID string, tagType shared.CardTag) int {
-	p, err := ctx.game.GetPlayer(playerID)
-	if err != nil {
-		return 0
-	}
-	count := 0
-	if ctx.game.vpCardLookup == nil {
-		return 0
-	}
-	for _, cardID := range p.PlayedCards().Cards() {
-		cardInfo, err := ctx.game.vpCardLookup.LookupVPCard(cardID)
-		if err != nil {
-			continue
-		}
-		if cardInfo.CardType == "event" && tagType != shared.TagEvent {
-			continue
-		}
-		for _, tag := range cardInfo.Tags {
-			if tag == tagType || tag == shared.TagWild {
-				count++
-			}
-		}
-	}
-	return count
-}
-
-func (ctx *gameVPRecalculationContext) CountAllTilesOfType(tileType shared.ResourceType) int {
-	tiles := ctx.game.board.Tiles()
-	count := 0
-	for _, tile := range tiles {
-		if tile.OccupiedBy != nil && tile.OccupiedBy.Type == tileType {
-			count++
-		}
-	}
-	return count
-}
-
-func (ctx *gameVPRecalculationContext) CountPlayerTilesOfType(playerID string, tileType shared.ResourceType) int {
-	tiles := ctx.game.board.Tiles()
-	count := 0
-	for _, tile := range tiles {
-		if tile.OccupiedBy != nil && tile.OccupiedBy.Type == tileType &&
-			tile.OwnerID != nil && *tile.OwnerID == playerID {
-			count++
-		}
-	}
-	return count
-}
-
-func (ctx *gameVPRecalculationContext) CountAdjacentTilesForCard(cardID string, tileType shared.ResourceType) int {
-	tiles := ctx.game.board.Tiles()
-	sourceTag := "source:" + cardID
-
-	var sourceTile *board.Tile
-	for i := range tiles {
-		if tiles[i].OccupiedBy == nil {
-			continue
-		}
-		if slices.Contains(tiles[i].OccupiedBy.Tags, sourceTag) {
-			sourceTile = &tiles[i]
-			break
-		}
-	}
-
-	if sourceTile == nil {
-		return 0
-	}
-
-	neighbors := sourceTile.Coordinates.GetNeighbors()
-	count := 0
-	for _, tile := range tiles {
-		if tile.OccupiedBy == nil || tile.OccupiedBy.Type != tileType {
-			continue
-		}
-		if slices.Contains(neighbors, tile.Coordinates) {
-			count++
-		}
-	}
-	return count
-}
-
-func (ctx *gameVPRecalculationContext) CountAdjacentTilesToTileType(playerID string, countType, adjacentToType shared.ResourceType) int {
-	tiles := ctx.game.board.Tiles()
-
-	tilesByCoord := make(map[shared.HexPosition]*board.Tile, len(tiles))
-	for i := range tiles {
-		tilesByCoord[tiles[i].Coordinates] = &tiles[i]
-	}
-
-	counted := make(map[shared.HexPosition]bool)
-
-	for _, tile := range tiles {
-		if tile.OwnerID == nil || *tile.OwnerID != playerID {
-			continue
-		}
-		if tile.OccupiedBy == nil || tile.OccupiedBy.Type != adjacentToType {
-			continue
-		}
-
-		for _, neighborCoord := range tile.Coordinates.GetNeighbors() {
-			neighborTile, exists := tilesByCoord[neighborCoord]
-			if !exists || neighborTile.OccupiedBy == nil || counted[neighborCoord] {
-				continue
-			}
-
-			var matches bool
-			if countType == shared.ResourceGreeneryTile {
-				matches = neighborTile.OccupiedBy.Type == shared.ResourceGreeneryTile || neighborTile.OccupiedBy.Type == shared.ResourceWorldTreeTile
-			} else {
-				matches = neighborTile.OccupiedBy.Type == countType
-			}
-
-			if matches {
-				counted[neighborCoord] = true
-			}
-		}
-	}
-
-	return len(counted)
-}
-
-func (ctx *gameVPRecalculationContext) CountAllColonies() int {
-	return ctx.game.Colonies().CountAllColonies()
-}
-
 type Game struct {
 	mu               sync.RWMutex
 	ds               *datastore.DataStore
@@ -1874,295 +1724,6 @@ func (g *Game) GetTriggeredEffects() []shared.TriggeredEffect {
 func (g *Game) ClearTriggeredEffects() {
 	g.update(func(s *datastore.GameState) {
 		s.TriggeredEffects = nil
-	})
-}
-
-func (g *Game) SetVPCardLookup(lookup VPCardLookup) {
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	g.vpCardLookup = lookup
-	g.subscribeToVPEvents()
-}
-
-// RegisterCorporationVPGranter registers VP conditions from a corporation card.
-func (g *Game) RegisterCorporationVPGranter(playerID string, corporationID string) {
-	if g.vpCardLookup == nil {
-		return
-	}
-	cardInfo, err := g.vpCardLookup.LookupVPCard(corporationID)
-	if err != nil || len(cardInfo.VPConditions) == 0 {
-		return
-	}
-	p, err := g.GetPlayer(playerID)
-	if err != nil {
-		return
-	}
-	granter := shared.VPGranter{
-		CardID:       cardInfo.CardID,
-		CardName:     cardInfo.CardName,
-		Description:  cardInfo.Description,
-		VPConditions: cardInfo.VPConditions,
-	}
-	p.VPGranters().Prepend(granter)
-	g.recalculatePlayerVP(p)
-}
-
-func (g *Game) recalculatePlayerVP(p *player.Player) {
-	if g.vpCardLookup == nil {
-		return
-	}
-	ctx := &gameVPRecalculationContext{game: g}
-	p.VPGranters().RecalculateAll(ctx)
-}
-
-func (g *Game) recalculateAllPlayersVP() {
-	for _, p := range g.GetAllPlayers() {
-		g.recalculatePlayerVP(p)
-	}
-}
-
-func (g *Game) subscribeToVPEvents() {
-	events.Subscribe(g.eventBus, func(e events.CardPlayedEvent) {
-		if g.vpCardLookup == nil {
-			return
-		}
-		cardInfo, err := g.vpCardLookup.LookupVPCard(e.CardID)
-		if err != nil {
-			return
-		}
-		if len(cardInfo.VPConditions) == 0 {
-			return
-		}
-
-		p, err := g.GetPlayer(e.PlayerID)
-		if err != nil {
-			return
-		}
-
-		granter := shared.VPGranter{
-			CardID:       cardInfo.CardID,
-			CardName:     cardInfo.CardName,
-			Description:  cardInfo.Description,
-			VPConditions: cardInfo.VPConditions,
-		}
-		p.VPGranters().Add(granter)
-		g.recalculatePlayerVP(p)
-	})
-
-	events.Subscribe(g.eventBus, func(e events.ResourceStorageChangedEvent) {
-		p, err := g.GetPlayer(e.PlayerID)
-		if err != nil {
-			return
-		}
-		g.recalculatePlayerVP(p)
-	})
-
-	events.Subscribe(g.eventBus, func(e events.TagPlayedEvent) {
-		p, err := g.GetPlayer(e.PlayerID)
-		if err != nil {
-			return
-		}
-		g.recalculatePlayerVP(p)
-	})
-
-	events.Subscribe(g.eventBus, func(e events.TilePlacedEvent) {
-		g.recalculateAllPlayersVP()
-	})
-
-	events.Subscribe(g.eventBus, func(e events.ColonyBuiltEvent) {
-		g.recalculateAllPlayersVP()
-	})
-
-}
-
-func (g *Game) subscribeToGenerationalEvents() {
-	events.Subscribe(g.eventBus, func(e events.TerraformRatingChangedEvent) {
-		if e.NewRating > e.OldRating {
-			p, err := g.GetPlayer(e.PlayerID)
-			if err != nil {
-				return
-			}
-			p.GenerationalEvents().Increment(shared.GenerationalEventTRRaise)
-		}
-	})
-
-	events.Subscribe(g.eventBus, func(e events.TilePlacedEvent) {
-		p, err := g.GetPlayer(e.PlayerID)
-		if err != nil {
-			return
-		}
-
-		switch e.TileType {
-		case "ocean":
-			p.GenerationalEvents().Increment(shared.GenerationalEventOceanPlacement)
-		case "city":
-			p.GenerationalEvents().Increment(shared.GenerationalEventCityPlacement)
-		case "greenery":
-			p.GenerationalEvents().Increment(shared.GenerationalEventGreeneryPlacement)
-		}
-	})
-
-	events.Subscribe(g.eventBus, func(e events.GenerationAdvancedEvent) {
-		for _, p := range g.GetAllPlayers() {
-			p.GenerationalEvents().Clear()
-			// Clear temporary "generation-end" effects
-			p.Effects().RemoveTemporaryEffects(shared.TemporaryGenerationEnd)
-			// Also clear any "next-card" effects that weren't consumed
-			p.Effects().RemoveTemporaryEffects(shared.TemporaryNextCard)
-		}
-	})
-}
-
-func (g *Game) subscribeToOceanSpaceEvents() {
-	events.Subscribe(g.eventBus, func(e events.TilePlacedEvent) {
-		if e.TileType == string(shared.ResourceOceanTile) || e.TileType == "ocean" {
-			return
-		}
-
-		coords := shared.HexPosition{Q: e.Q, R: e.R, S: e.S}
-		tile, err := g.board.GetTile(coords)
-		if err != nil {
-			return
-		}
-		if tile.Type != shared.ResourceOceanSpace {
-			return
-		}
-
-		freeOceanSpaces := g.board.FreeOceanSpaces()
-		gp := g.globalParameters
-		oceansRemaining := gp.GetMaxOceans() - gp.Oceans()
-		if freeOceanSpaces < oceansRemaining {
-			gp.ReduceMaxOceans(gp.Oceans() + freeOceanSpaces)
-		}
-	})
-}
-
-func (g *Game) subscribeToGlobalParameterBonuses() {
-	log := slog.Default()
-
-	events.Subscribe(g.eventBus, func(e events.TemperatureChangedEvent) {
-		if e.ChangedBy == "" {
-			return
-		}
-		p, err := g.GetPlayer(e.ChangedBy)
-		if err != nil {
-			return
-		}
-
-		if e.OldValue < -24 && e.NewValue >= -24 {
-			p.Resources().AddProduction(map[shared.ResourceType]int{
-				shared.ResourceHeatProduction: 1,
-			})
-			g.AddTriggeredEffect(shared.TriggeredEffect{
-				CardName:   "Temperature Bonus",
-				PlayerID:   e.ChangedBy,
-				SourceType: shared.SourceTypeGlobalBonus,
-				CalculatedOutputs: []shared.CalculatedOutput{
-					{ResourceType: string(shared.ResourceHeatProduction), Amount: 1},
-				},
-			})
-			log.Debug("Temperature bonus: +1 heat production at -24C", slog.String("player_id", e.ChangedBy))
-		}
-
-		if e.OldValue < -20 && e.NewValue >= -20 {
-			p.Resources().AddProduction(map[shared.ResourceType]int{
-				shared.ResourceHeatProduction: 1,
-			})
-			g.AddTriggeredEffect(shared.TriggeredEffect{
-				CardName:   "Temperature Bonus",
-				PlayerID:   e.ChangedBy,
-				SourceType: shared.SourceTypeGlobalBonus,
-				CalculatedOutputs: []shared.CalculatedOutput{
-					{ResourceType: string(shared.ResourceHeatProduction), Amount: 1},
-				},
-			})
-			log.Debug("Temperature bonus: +1 heat production at -20C", slog.String("player_id", e.ChangedBy))
-		}
-
-		if e.OldValue < 0 && e.NewValue >= 0 {
-			ctx := context.Background()
-			if err := g.AppendToPendingTileSelectionQueue(ctx, e.ChangedBy, []string{"ocean"}, "Temperature Bonus", "", nil); err != nil {
-				log.Warn("Failed to queue ocean tile for temperature bonus", slog.Any("error", err))
-			}
-			g.AddTriggeredEffect(shared.TriggeredEffect{
-				CardName:   "Temperature Bonus",
-				PlayerID:   e.ChangedBy,
-				SourceType: shared.SourceTypeGlobalBonus,
-				CalculatedOutputs: []shared.CalculatedOutput{
-					{ResourceType: string(shared.ResourceOceanTile), Amount: 1},
-				},
-			})
-			log.Debug("Temperature bonus: place ocean at 0C", slog.String("player_id", e.ChangedBy))
-		}
-	})
-
-	events.Subscribe(g.eventBus, func(e events.OxygenChangedEvent) {
-		if e.ChangedBy == "" {
-			return
-		}
-
-		if e.OldValue < 8 && e.NewValue >= 8 {
-			ctx := context.Background()
-			actualSteps, err := g.globalParameters.IncreaseTemperature(ctx, 1, e.ChangedBy)
-			if err != nil {
-				slog.Default().Warn("Failed to increase temperature from oxygen bonus", slog.Any("error", err))
-			}
-			if actualSteps > 0 {
-				p, pErr := g.GetPlayer(e.ChangedBy)
-				if pErr == nil {
-					p.Resources().UpdateTerraformRating(actualSteps)
-				}
-			}
-			g.AddTriggeredEffect(shared.TriggeredEffect{
-				CardName:   "Oxygen Bonus",
-				PlayerID:   e.ChangedBy,
-				SourceType: shared.SourceTypeGlobalBonus,
-				CalculatedOutputs: []shared.CalculatedOutput{
-					{ResourceType: string(shared.ResourceTemperature), Amount: 1},
-				},
-			})
-			log.Debug("Oxygen bonus: +1 temperature step at 8%", slog.String("player_id", e.ChangedBy))
-		}
-	})
-
-	events.Subscribe(g.eventBus, func(e events.VenusChangedEvent) {
-		if e.ChangedBy == "" {
-			return
-		}
-		p, err := g.GetPlayer(e.ChangedBy)
-		if err != nil {
-			return
-		}
-
-		if e.OldValue < 8 && e.NewValue >= 8 {
-			ctx := context.Background()
-			cardIDs, drawErr := g.deck.DrawProjectCards(ctx, 1)
-			if drawErr == nil && len(cardIDs) > 0 {
-				p.Hand().AddCard(cardIDs[0])
-			}
-			g.AddTriggeredEffect(shared.TriggeredEffect{
-				CardName:   "Venus Bonus",
-				PlayerID:   e.ChangedBy,
-				SourceType: shared.SourceTypeGlobalBonus,
-				CalculatedOutputs: []shared.CalculatedOutput{
-					{ResourceType: "card-draw", Amount: 1},
-				},
-			})
-			log.Debug("Venus bonus: draw 1 card at 8%", slog.String("player_id", e.ChangedBy))
-		}
-
-		if e.OldValue < 16 && e.NewValue >= 16 {
-			p.Resources().UpdateTerraformRating(1)
-			g.AddTriggeredEffect(shared.TriggeredEffect{
-				CardName:   "Venus Bonus",
-				PlayerID:   e.ChangedBy,
-				SourceType: shared.SourceTypeGlobalBonus,
-				CalculatedOutputs: []shared.CalculatedOutput{
-					{ResourceType: string(shared.ResourceTR), Amount: 1},
-				},
-			})
-			log.Debug("Venus bonus: +1 TR at 16%", slog.String("player_id", e.ChangedBy))
-		}
 	})
 }
